+++ seacatauth/external_login/providers/generic.py
@@ -12,6 +12,8 @@ import jwcrypto.jwt
 import jwcrypto.jwk
 import jwcrypto.jws
 
+from ..exceptions import ExternalOAuthFlowError
+
 #
 
 L = logging.getLogger(__name__)
@@ -40,6 +42,11 @@ class GenericOAuth2Login(asab.Configurable):
 	scope=openid name email
 	label=Login via provider.auth
 	```
+
+	Seacat Auth external login callback endpoint (/public/ext-login/callback) must be allowed as a redirect URIs
+	in the OAuth client settings at the external login account provider.
+	The full callback URL is canonically in the following format:
+	https://{my_domain}/api/seacat-auth/public/ext-login/callback
 	"""
 
 	Type = None
@@ -89,36 +96,36 @@ class GenericOAuth2Login(asab.Configurable):
 		self.Ident = self.Config.get("ident", "email")
 		assert self.Ident is not None
 
-		self.IDClaimsToVerify = {
-			"iss": self.Issuer,
-			"aud": self.ClientId
-		}
-
 		# Label for "Sign up with {ext_login_provider}" button
 		# TODO: Make this i18n-compatible (like login descriptors)
 		# TODO: Separate label for "Add external login" button
 		self.Label = self.Config.get("label")
 		assert self.Label is not None
 
-		# Base URL under which the external login endpoint are available
-		public_api_base_url = self.Config.get("public_api_base_url")
-		if public_api_base_url is None:
-			# Fallback to general public_api_base_url
-			public_api_base_url = external_login_svc.App.PublicSeacatAuthApiUrl
-
 		self.JwkSet = None
 
-		self.LoginURI = "{}{}".format(
-			public_api_base_url.rstrip("/"),
-			external_login_svc.ExternalLoginPath.format(ext_login_provider=self.Type)
-		)
-		self.AddExternalLoginURI = "{}{}".format(
-			public_api_base_url.rstrip("/"),
-			external_login_svc.AddExternalLoginPath.format(ext_login_provider=self.Type)
-		)
+		# The URL to return to after successful external login
+		# Mostly for debugging purposes
+		if "_callback_url" in self.Config:
+			self.CallbackUrl = self.Config.get("_callback_url")
+		else:
+			self.CallbackUrl = external_login_svc.CallbackUrlTemplate.format(provider_type=self.Type)
+
 
 	async def initialize(self, app):
 		await self._prepare_jwks()
+
+
+	def acr_value(self) -> str:
+		"""
+		Authentication Context Class Reference (ACR)
+		https://openid.net/specs/openid-connect-core-1_0.html#AuthRequest
+
+		OpenID Connect clients may use an ACR value in the authorization request to specifically request which external
+		login provider should be used for End-User authentication.
+		"""
+		return "ext:{}".format(self.Type)
+
 
 	async def _prepare_jwks(self, speculative=True):
 		if not self.JwksUri:
@@ -141,8 +148,8 @@ class GenericOAuth2Login(asab.Configurable):
 		self.JwkSet = jwcrypto.jwk.JWKSet.from_json(jwks)
 		L.info("Identity provider public JWK set loaded.", struct_data={"type": self.Type})
 
-	def _get_authorize_uri(
-		self, redirect_uri: str,
+	def get_authorize_uri(
+		self, redirect_uri: typing.Optional[str] = None,
 		state: typing.Optional[str] = None,
 		nonce: typing.Optional[str] = None
 	):
@@ -150,7 +157,7 @@ class GenericOAuth2Login(asab.Configurable):
 			("response_type", "code"),
 			("client_id", self.ClientId),
 			("scope", self.Scope),
-			("redirect_uri", redirect_uri),
+			("redirect_uri", redirect_uri or self.CallbackUrl),
 			("prompt", "select_account"),
 		]
 		if state is not None:
@@ -163,7 +170,7 @@ class GenericOAuth2Login(asab.Configurable):
 		)
 
 	@contextlib.asynccontextmanager
-	async def token_request(self, code: str, redirect_uri: str):
+	async def token_request(self, code: str, redirect_uri: str | None = None):
 		"""
 		Send auth code to token request endpoint and return access token
 		"""
@@ -171,7 +178,7 @@ class GenericOAuth2Login(asab.Configurable):
 			("grant_type", "authorization_code"),
 			("code", code),
 			("client_id", self.ClientId),
-			("redirect_uri", redirect_uri)]
+			("redirect_uri", redirect_uri or self.CallbackUrl)]
 		if self.ClientSecret:
 			request_params.append(("client_secret", self.ClientSecret))
 		query_string = urllib.parse.urlencode(request_params)
@@ -183,16 +190,16 @@ class GenericOAuth2Login(asab.Configurable):
 			async with session.post(self.TokenEndpoint, data=query_string, headers=headers) as resp:
 				if resp.status != 200:
 					text = await resp.text()
-					L.error("Error response from external auth provider", struct_data={
+					L.error("Error response from external auth provider.", struct_data={
 						"status": resp.status,
 						"url": resp.url,
 						"text": text
 					})
-					yield None
+					raise ExternalOAuthFlowError("Token request failed.")
 				else:
 					yield resp
 
-	async def _get_user_info(self, authorize_data: dict, redirect_uri: str):
+	async def get_user_info(self, authorize_data: dict, expected_nonce: str | None = None) -> typing.Optional[dict]:
 		"""
 		Obtain the authenticated user's profile info, with the claims normalized to be in line with
 		OpenID UserInfo response.
@@ -211,26 +218,25 @@ class GenericOAuth2Login(asab.Configurable):
 			L.error("Code parameter not provided in authorize response.", struct_data={
 				"provider": self.Type,
 				"query": dict(authorize_data)})
-			return None
+			raise ExternalOAuthFlowError("No 'code' parameter in request.")
 
-		async with self.token_request(code, redirect_uri=redirect_uri) as resp:
-			if resp is None:
-				return None
+		async with self.token_request(code) as resp:
 			token_data = await resp.json()
 
 		if "id_token" not in token_data:
 			L.error("Token response does not contain 'id_token'", struct_data={
 				"provider": self.Type, "resp": token_data})
-			return None
+			raise ExternalOAuthFlowError("No 'id_token' in token response.")
 
 		id_token = token_data["id_token"]
 		await self._prepare_jwks()
 
-		id_token_claims = self._get_verified_claims(id_token)
-		user_info = await self._user_data_from_id_token_claims(id_token_claims)
+		id_token_claims = self._get_verified_claims(id_token, expected_nonce)
+		user_info = self._user_data_from_id_token_claims(id_token_claims)
+		user_info["sub"] = str(user_info["sub"])
 		return user_info
 
-	async def _user_data_from_id_token_claims(self, id_token_claims: dict):
+	def _user_data_from_id_token_claims(self, id_token_claims: dict):
 		user_info = {
 			k: v
 			for k, v in id_token_claims.items()
@@ -241,30 +247,28 @@ class GenericOAuth2Login(asab.Configurable):
 		}
 		return user_info
 
-	def _get_verified_claims(self, id_token):
+	def _get_verified_claims(self, id_token, expected_nonce: str | None = None):
+		check_claims = self._get_claims_to_verify()
+		if expected_nonce:
+			check_claims["nonce"] = expected_nonce
 		try:
-			id_token = jwcrypto.jwt.JWT(jwt=id_token, key=self.JwkSet, check_claims=self.IDClaimsToVerify)
+			id_token = jwcrypto.jwt.JWT(jwt=id_token, key=self.JwkSet, check_claims=check_claims)
 			claims = json.loads(id_token.claims)
 		except jwcrypto.jws.InvalidJWSSignature:
 			L.error("Invalid ID token signature.", struct_data={"provider": self.Type})
-			return None
+			raise ExternalOAuthFlowError("Invalid ID token signature.")
 		except jwcrypto.jwt.JWTExpired:
 			L.error("Expired ID token.", struct_data={"provider": self.Type})
-			return None
+			raise ExternalOAuthFlowError("Expired ID token.")
 		except Exception as e:
 			L.error("Error reading ID token claims.", struct_data={
 				"provider": self.Type, "error": str(e)})
-			return None
+			raise ExternalOAuthFlowError("Error reading ID token claims.")
 		return claims
 
-	def get_login_authorize_uri(self, state: typing.Optional[str] = None):
-		return self._get_authorize_uri(self.LoginURI, state)
 
-	def get_addlogin_authorize_uri(self, state: typing.Optional[str] = None):
-		return self._get_authorize_uri(self.AddExternalLoginURI, state)
-
-	async def do_external_login(self, authorize_data: dict):
-		return await self._get_user_info(authorize_data, redirect_uri=self.LoginURI)
-
-	async def add_external_login(self, authorize_data: dict):
-		return await self._get_user_info(authorize_data, redirect_uri=self.AddExternalLoginURI)
+	def _get_claims_to_verify(self) -> dict:
+		return {
+			"iss": self.Issuer,
+			"aud": self.ClientId
+		}