+++ seacatauth/session/adapter.py
@@ -66,6 +66,7 @@ class OAuth2Data:
 	ClientId: typing.Optional[str]
 	Scope: typing.Optional[str]
 	Nonce: typing.Optional[str]
+	RedirectUri: typing.Optional[str]
 
 
 @dataclasses.dataclass
@@ -141,6 +142,7 @@ class SessionAdapter:
 			Scope = "oa_sc"
 			ClientId = "oa_cl"
 			Nonce = "oa_no"
+			RedirectUri = "oa_ru"
 
 		class Cookie:
 			_prefix = "ck"
@@ -263,6 +265,7 @@ class SessionAdapter:
 				self.FN.OAuth2.ClientId: self.OAuth2.ClientId,
 				self.FN.OAuth2.Scope: self.OAuth2.Scope,
 				self.FN.OAuth2.Nonce: self.OAuth2.Nonce,
+				self.FN.OAuth2.RedirectUri: self.OAuth2.RedirectUri,
 			})
 
 		if self.Batman is not None:
@@ -430,6 +433,7 @@ class SessionAdapter:
 			Scope=session_dict.pop(cls.FN.OAuth2.Scope, None),
 			ClientId=session_dict.pop(cls.FN.OAuth2.ClientId, None),
 			Nonce=session_dict.pop(cls.FN.OAuth2.Nonce, None),
+			RedirectUri=session_dict.pop(cls.FN.OAuth2.RedirectUri, None),
 		)
 
 	@classmethod